Updated the script to only print failing test by default

+++ tests/run-tests.sh
@@ -5,55 +5,103 @@ failed_count=0
 GENERATOR_COUNT=20
 
 function runTest {
+    if [ -z "$PRINT_ALL" ]
+    then
+        if [ $overwrite ]
+        then
+            echo -n -e "\e[3A"
+        else
+            overwrite=1
+        fi
+        echo
+        echo -e "\e[32mPassed $passed_count tests\e[m"
+        echo -e "\e[31mFailed $failed_count tests\e[m"
+    fi
     if [ ${1: -6} == .basic ]
     then
-        for i in $(seq 1 $2)
-        do
-            echo -n "  "
-        done
+        if [ $PRINT_ALL ]
+        then
+            for i in $(seq 1 $2)
+            do
+                echo -n "  "
+            done
+        fi
         if timeout 1s $COMPILER $1 &> /tmp/basic-test.out
         then
             if [ ${3::5} == fail. ]
             then
+                if [ -z "$PRINT_ALL" ]
+                then
+                    echo -n -e "\e[3A\e[J"
+                    unset overwrite
+                fi
                 echo -e "\e[31mFailed\e[m test '$3' should have failed"
                 failed_count=$(expr $failed_count + 1)
+                overwrite=""
             else 
-                echo -e "\e[32mPassed\e[m test '$3'"
                 passed_count=$(expr $passed_count + 1)
+                if [ $PRINT_ALL ]
+                then
+                    echo -e "\e[32mPassed\e[m test '$3'"
+                fi
             fi
         else
             if [ $? == 124 ]
             then
                 if [ ${3::5} == fail. ]
                 then
-                    echo -e "\e[32mPassed\e[m test '$3'"
                     passed_count=$(expr $passed_count + 1)
+                    if [ $PRINT_ALL ]
+                    then
+                        echo -e "\e[32mPassed\e[m test '$3'"
+                    fi
                 else 
+                    if [ -z "$PRINT_ALL" ]
+                    then
+                        echo -n -e "\e[3A\e[J"
+                        unset overwrite
+                    fi
                     echo -e "\e[31mFailed\e[m test '$3' by timeout"
                     failed_count=$(expr $failed_count + 1)
+                    overwrite=""
                 fi
             else
                 if [ ${3::5} == fail. ]
                 then
-                    echo -e "\e[32mPassed\e[m test '$3'"
                     passed_count=$(expr $passed_count + 1)
+                    if [ $PRINT_ALL ]
+                    then
+                        echo -e "\e[32mPassed\e[m test '$3'"
+                    fi
                 else 
+                    if [ -z "$PRINT_ALL" ]
+                    then
+                        echo -n -e "\e[3A\e[J"
+                        unset overwrite
+                    fi
                     echo -e "\e[31mFailed\e[m test '$3' at runtime"
                     failed_count=$(expr $failed_count + 1)
-                    for i in $(seq 0 $2)
-                    do
-                        sed -i "s/^/  /" /tmp/basic-test.out
-                    done
+                    if [ $PRINT_ALL ]
+                    then
+                        for i in $(seq 0 $2)
+                        do
+                            sed -i "s/^/  /" /tmp/basic-test.out
+                        done
+                    fi
                     cat /tmp/basic-test.out
+                    overwrite=""
                 fi
             fi            
         fi
     elif [ ${1: -3} == .js -a -z "$NO_GEN" ]
     then
-        for i in $(seq 1 $2)
-        do
-            echo -n "  "
-        done
+        if [ $PRINT_ALL ]
+        then
+            for i in $(seq 1 $2)
+            do
+                echo -n "  "
+            done
+        fi
         dir=$(dirname $1)
         tmp_file=$dir/tmp.$3.basic
         unset failed
@@ -64,6 +112,11 @@ function runTest {
             then
                 if [ ${3::5} == fail. ]
                 then
+                    if [ -z "$PRINT_ALL" ]
+                    then
+                        echo -n -e "\e[3A\e[J"
+                        unset overwrite
+                    fi
                     echo -e "\e[31mFailed\e[m test '$3' should have failed"
                     failed=True
                     break
@@ -73,6 +126,11 @@ function runTest {
                 then
                     if [ ! ${3::5} == fail. ]
                     then
+                        if [ -z "$PRINT_ALL" ]
+                        then
+                            echo -n -e "\e[3A\e[J"
+                            unset overwrite
+                        fi
                         echo -e "\e[31mFailed\e[m test '$3' by timeout"
                         failed=True
                         break
@@ -80,12 +138,20 @@ function runTest {
                 else
                     if [ ! ${3::5} == fail. ]
                     then
+                        if [ -z "$PRINT_ALL" ]
+                        then
+                            echo -n -e "\e[3A\e[J"
+                            unset overwrite
+                        fi
                         echo -e "\e[31mFailed\e[m test '$3' at runtime"
                         failed=True
-                        for i in $(seq 0 $2)
-                        do
-                            sed -i "s/^/  /" /tmp/basic-test.out
-                        done
+                        if [ $PRINT_ALL ]
+                        then
+                            for i in $(seq 0 $2)
+                            do
+                                sed -i "s/^/  /" /tmp/basic-test.out
+                            done
+                        fi
                         cat /tmp/basic-test.out
                         break
                     fi
@@ -95,20 +161,27 @@ function runTest {
         if [ $failed ]
         then
             failed_count=$(expr $failed_count + 1)
+            overwrite=""
         else 
-            echo -e "\e[32mPassed\e[m test '$3'"
             passed_count=$(expr $passed_count + 1)
-            rm $tmp_file
+            if [ $PRINT_ALL ]
+            then
+                echo -e "\e[32mPassed\e[m test '$3'"
+                rm $tmp_file
+            fi
         fi
     fi
 }
 
 function runTests {
-    for i in $(seq 1 $2)
-    do
-        echo -n "  "
-    done
-    echo "Running tests in: '$3'"
+    if [ $PRINT_ALL ]
+    then
+        for i in $(seq 1 $2)
+        do
+            echo -n "  "
+        done
+        echo "Running tests in: '$3'"
+    fi
     for test in $(find $1 -mindepth 1 -maxdepth 1)
     do
         name=$(awk -F/ '{print $NF}' <<< $test)
@@ -133,10 +206,15 @@ then
     shift
     COMPILER=$@
     runTests $path 1 $name
-    echo
 
+    if [ $overwrite ]
+    then
+        echo -n -e "\e[3A"
+    fi
+    echo
     echo -e "\e[32mPassed $passed_count tests\e[m"
     echo -e "\e[31mFailed $failed_count tests\e[m"
+    echo
     if [ $failed_count -gt 0 ]
     then
         exit 1
